[add] function/macro experiments
I spend a whole morning playing with this and found some interesting
things.

First, you cannot assign functions to module attributes UNLESS they are
of the &mod.fun/arity variety. This means I HAVE to switch to a process-based
accumulator.

The good news is that my accumulator probably already hides all the
importand implementation details, so the switch should be pretty easy to
make.